Mv api/ to main dir. Mv main.go & cmds/ to adlrtool
