+++ Makefile
@@ -10,13 +10,14 @@ INTEGRATION=internal/integration
 TIMEOUT=5m
 
 ADLR_SRC=pkg
+ADLR_MAIN=adlrtool
 MOCK=mockery
 MOCKS=internal/mocks
 
 BIN=bin
 SCRIPTS=sh
 
-LICENSELOCK=license.lock
+LICENSELOCK=adlrtool/license.lock
 BUILDLIST=buildlist.json
 
 default: test
@@ -41,18 +42,20 @@ bin:
 build: build-linux-amd64
 
 build-tmp: bin # build tmp exec to perform adlr steps
-	@$(GOBUILD) -o $(BIN)/tmp .
+	@$(GOBUILD) -o $(BIN)/tmp ./$(ADLR_MAIN)
 
 build-linux-amd64: licenselock
 	@$(SCRIPTS)/build.sh \
 	adlr linux amd64 \
-	. ./$(BIN) ./$(LICENSELOCK)
+	./$(ADLR_MAIN) ./$(BIN) ./$(LICENSELOCK)
 
 buildlist:
 	@$(GOLIST) -m -json all > $(BUILDLIST)
 
 licenselock: build-tmp buildlist
-	@$(BIN)/tmp evaluate --buildlist=$(BUILDLIST)
+	@$(BIN)/tmp evaluate \
+	--buildlist=$(BUILDLIST) \
+	--dir=$(ADLR_MAIN)
 
 # testing
 test: test-unit test-integration
@@ -62,4 +65,4 @@ test-integration: buildlist
 	&& $(GOTIDY)
 
 test-unit:
-	@$(GOTEST) -short ./... && $(GOTIDY)
+	@$(GOTEST) -short ./$(ADLR_SRC)/... && $(GOTIDY)
