+++ README.md
@@ -51,8 +51,8 @@ Unfortunately, golang does not yet have a standard for module license files. The
 ```golang
 direct := gotool.FilterDirectImportModules(mods)
 
-prospects := api.MakeProspects(direct...)
-prospector := api.MakeProspector()
+prospects := adlr.MakeProspects(direct...)
+prospector := adlr.MakeProspector()
 mines, err := prospector.Prospect(prospects...)
 ...
 ```
@@ -63,7 +63,7 @@ From prospecting, one or multiple matches are returned for a golang module with 
 mines, err := prospector.Prospect(prospects...)
 ...
 
-miner := api.MakeMiner()
+miner := adlr.MakeMiner()
 locks, err := miner.Mine(mines...)
 if err != nil && Verbose {
 	fmt.Println(err)
@@ -78,7 +78,7 @@ New dependencies take priority, and will fill the lock file. But for new locks t
 locks, err := miner.Mine(mines...)
 ...
 
-licenselock := api.MakeLicenseLockManager("./")
+licenselock := adlr.MakeLicenseLockManager("./")
 err = licenselock.Lock(locks...)
 ...
 ```
@@ -86,15 +86,15 @@ err = licenselock.Lock(locks...)
 ## Auditing Locked License types
 After locking, dependencies and their licenses have been written to the lock file. But unwanted license types may have slipped through. The auditing step will search through the lock file, checking license types against a whitelist. For any types not listed, an error is returned listing bad license types, and requesting whitelist inclusion or dependency removal.
 ```golang
-licenselock := api.MakeLicenseLockManager("./")
+licenselock := adlr.MakeLicenseLockManager("./")
 err = licenselock.Lock(locks...)
 ...
 
 locks, err = licenselock.Read()
 ...
 
-whitelist := api.MakeWhitelist([]string{"A","B","C"...})
-auditor := api.MakeAuditor(whitelist)
+whitelist := adlr.MakeWhitelist([]string{"A","B","C"...})
+auditor := adlr.MakeAuditor(whitelist)
 err = auditor.Audit(locks...)
 ...
 ```
