+++ adlrtool/cmd/evaluate.go
@@ -5,7 +5,7 @@ import (
 
 	"github.com/spf13/cobra"
 
-	"github.com/blocky/adlr/api"
+	"github.com/blocky/adlr"
 	"github.com/blocky/adlr/pkg/gotool"
 	"github.com/blocky/prettyprinter"
 )
@@ -50,27 +50,27 @@ func Evaluate(buildlist *os.File) {
 	ExitOnErr(err)
 
 	direct := gotool.FilterDirectImportModules(mods)
-	prospects := api.MakeProspects(direct...)
+	prospects := adlr.MakeProspects(direct...)
 
-	prospector := api.MakeProspector()
+	prospector := adlr.MakeProspector()
 	mines, err := prospector.Prospect(prospects...)
 	ExitOnErr(err)
 
-	miner := api.MakeMiner()
+	miner := adlr.MakeMiner()
 	locks, err := miner.Mine(mines...)
 	if Verbose && err != nil {
 		PrintStderr(err)
 	}
 
-	licenselock := api.MakeLicenseLockManager(ModuleDir)
+	licenselock := adlr.MakeLicenseLockManager(ModuleDir)
 	err = licenselock.Lock(locks...)
 	ExitOnErr(err)
 
 	locks, err = licenselock.Read()
 	ExitOnErr(err)
 
-	whitelist := api.MakeWhitelist(api.DefaultWhitelist) //https://github.com/blocky/adlr/issues/23
-	auditor := api.MakeAuditor(whitelist)
+	whitelist := adlr.MakeWhitelist(adlr.DefaultWhitelist) //https://github.com/blocky/adlr/issues/23
+	auditor := adlr.MakeAuditor(whitelist)
 	err = auditor.Audit(locks...)
 	ExitOnErr(err)
 }
