+++ adlrtool/license.lock
