--- api/license.go
@@ -1,6 +1,0 @@
-package api
-
-import "github.com/blocky/adlr/pkg/ascertain"
-
-// License is a type alias for ascertain.License
-type License = ascertain.License
