--- api/lock.go
@@ -1,21 +1,0 @@
-package api
-
-import "github.com/blocky/adlr/pkg/ascertain"
-
-// DependencyLock is a type alias for ascertain.DependencyLock
-type DependencyLock = ascertain.DependencyLock
-
-// DepLocksToDepLockMap is function alias for ascertain.DepLocksToDepLockMap
-var DepLocksToDepLockMap = ascertain.DepLocksToDepLockMap
-
-// MarshalDependencyLocks is function alias for ascertain.MarshalDependencyLocks
-var MarshalDependencyLocks = ascertain.MarshalDependencyLocks
-
-// UnmarshalDependencyLocks is a function alias for ascertain.UnmarshalDependencyLocks
-var UnmarshalDependencyLocks = ascertain.UnmarshalDependencyLocks
-
-// DeserializeLocks is function alias for ascertain.DeserializeLocks
-var DeserializeLocks = ascertain.DeserializeLocks
-
-// SerializeLocks is function alias for ascertain.SerializeLocks
-var SerializeLocks = ascertain.SerializeLocks
