--- api/mine.go
@@ -1,6 +1,0 @@
-package api
-
-import "github.com/blocky/adlr/pkg/ascertain"
-
-// Mine is a type alias for ascertain.Mine
-type Mine = ascertain.Mine
