--- api/prospect.go
@@ -1,9 +1,0 @@
-package api
-
-import "github.com/blocky/adlr/pkg/ascertain"
-
-// Prospect is a type alias for ascertain.Prospect
-type Prospect = ascertain.Prospect
-
-// MakeProspects is a type alias for ascertain.MakeProspects
-var MakeProspects = ascertain.MakeProspects
